* CRM-287 Fixed Rollbacks to Text Stop Updates

+++ database/migrations/2020_08_18_225845_add_crm_text_stop_sms_number.php
@@ -14,15 +14,19 @@ class AddCrmTextStopSmsNumber extends Migration
     public function up()
     {
         Schema::table('crm_text_stop', function (Blueprint $table) {
-            $table->string('sms_number', 12)->index();
+            if(Schema::hasColumn('crm_text_stop', 'sms_number')) {
+                $table->string('sms_number', 12)->index();
+            }
 
-            $table->dropColumn('response_id');
+            if(Schema::hasColumn('crm_text_stop', 'response_id')) {
+                $table->dropColumn('response_id');
 
-            /*$table->dropForeign('crm_text_stop_lead_id_foreign');
+                $table->dropForeign('crm_text_stop_lead_id_foreign');
 
-            $table->dropForeign('crm_text_stop_text_id_foreign');
+                $table->dropForeign('crm_text_stop_text_id_foreign');
 
-            $table->dropForeign('crm_text_stop_response_id_foreign');*/
+                $table->dropForeign('crm_text_stop_response_id_foreign');
+            }
         });
     }
 
@@ -34,21 +38,7 @@ class AddCrmTextStopSmsNumber extends Migration
     public function down()
     {
         Schema::table('crm_text_stop', function (Blueprint $table) {
-            $table->dropColumn('sms_number');
-
-            $table->integer('response_id');
-
-            $table->foreign('lead_id')
-                    ->references('identifier')
-                    ->on('website_lead');
-
-            $table->foreign('text_id')
-                    ->references('id')
-                    ->on('dealer_texts_log');
-
-            $table->foreign('response_id')
-                    ->references('id')
-                    ->on('dealer_texts_log');
+            //
         });
     }
 }
